src/stac_ingestion/generate_stac_item/requirements.txt
affine==2.3.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
azure-common==1.1.28
azure-core==1.25.0
azure-identity==1.10.0
azure-servicebus==7.8.0
azure-storage-blob==12.13.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cryptography==37.0.4
frozenlist==1.3.1
fsspec==2022.8.0
GDAL==3.5.1
google-api-core==2.8.2
google-auth==2.11.0
googleapis-common-protos==1.56.4
idna==3.3
importlib-resources==5.9.0
iniconfig==1.1.1
isodate==0.6.1
jsonschema==4.14.0
lxml==4.9.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.2
numpy==1.23.2
oauthlib==3.2.0
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
packaging==21.3
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
portalocker==2.5.1
protobuf==4.21.5
psutil==5.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyproj==3.3.1
pyrsistent==0.18.1
pystac==1.6.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
rasterio==1.3.0.post1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Shapely==1.8.4
six==1.16.0
snuggs==1.4.7
stac-check==1.2.0
stac-validator==3.1.0
stactools==0.4.1
tomli==2.0.1
types-setuptools==65.3.0
typing_extensions==4.3.0
uamqp==1.6.0
urllib3==1.26.12
yarl==1.8.1
zipp==3.8.1
